:truck: [#1929] -- move upload validator and tests into formio package
